requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pytest
psycopg2